Cambiando algunos floats por doubles...

+++ GUI_SI/src/imagen/ImageClass.java
@@ -455,7 +455,7 @@ public class ImageClass {
         for (int i=0;i<256;i++){
             trans_table[i] = Math.max(0, Math.round((256/img_size)*acHisto[i])-1);
             //System.out.println(trans_table[i]);
-            trans_table[i] = (int) Math.max(0, Math.round(((float)256/(float)img_size)*acHisto[i])-1);
+            trans_table[i] = (int) Math.max(0, Math.round(((double)256/(double)img_size)*acHisto[i])-1);
             //System.out.println(trans_table[i]);
         }
         return TRANSFORM(pixels,trans_table);
